R: move SpecificSymbolValueSource to its own file
